requirements.txt
fastapi
uvicorn
python-dotenv
pydantic[email]
sqlalchemy
jose
passlib[bcrypt]
email-validator
python-multipart